requirements.txt
twilio==6.62.0
Django==3.2.5
djangorestframework==3.12.4
PyJWT==2.1.0
whitenoise==5.3.0
wincertstore==0.2
gunicorn==20.0.4
mysqlclient==2.0.1
